Update dupe detection
