+++ src/main/java/seedu/duke/exceptions/ExceptionType.java
@@ -11,7 +11,6 @@ public enum ExceptionType {
     INVALID_FACULTY("Oh no! Please enter a faculty."),
     INVALID_FOODPLACE("Oh no! I cannot detect any inputs."),
     INVALID_FAVOURITE("Oh no! I cannot seem to detect the command you want to store as your favourite."),
-    DUPLICATE_FAVOURITE("Oh no! This command is already saved as your favourite."),
     UPDATE_FILE_FAIL("Sorry to inform you that the update of frequency list is unsuccessful."),
     READ_FILE_FAIL("It is unfortunate that the program is unable to read the file");
