+++ src/main/java/seedu/duke/favorite/Fav.java
@@ -1,5 +1,7 @@
 package seedu.duke.favorite;
 
+import static seedu.duke.ui.Ui.printDupeMessage;
+
 public class Fav {
     private String command;
     private String desc;
@@ -9,8 +11,13 @@ public class Fav {
         this.desc = desc;
     }
 
-    public Boolean equals(Fav item) {
-        return this.command.equals(item.command) && this.desc.equals(item.desc);
+    public Boolean equals(Fav item, int index) {
+        assert item != null : "Invalid user favourites!";
+        if (this.command.equals(item.command) || this.desc.equals(item.desc)) {
+            printDupeMessage(index, this.desc, this.command);
+            return true;
+        }
+        return false;
     }
 
     public String getCommand() {
