+++ src/main/java/seedu/duke/logic/parser/Parser.java
@@ -2,7 +2,6 @@ package seedu.duke.logic.parser;
 
 import seedu.duke.exceptions.CustomException;
 import seedu.duke.exceptions.ExceptionType;
-import seedu.duke.logic.commands.AddFavCommand;
 import seedu.duke.logic.commands.AllBusCommand;
 import seedu.duke.logic.commands.BusCommand;
 import seedu.duke.logic.commands.DineCommand;
@@ -14,6 +13,7 @@ import seedu.duke.logic.commands.ListFavCommand;
 import seedu.duke.logic.commands.ListStopsCommand;
 import seedu.duke.logic.commands.RouteCommand;
 import seedu.duke.logic.commands.RouteMapCommand;
+import seedu.duke.logic.commands.favcommand.AddFavCommand;
 import seedu.duke.logic.commands.favcommand.DeleteFavCommand;
 import seedu.duke.logic.commands.favcommand.DescFavCommand;
 import seedu.duke.logic.commands.favcommand.ExecFavCommand;
